+++ src/custom/state/swap/TradeGp.ts
@@ -1,5 +1,5 @@
 import { CanonicalMarketParams, getCanonicalMarket } from 'utils/misc'
-import { FeeInformation, PriceInformation } from 'utils/price'
+import { FeeInformation, PriceInformation } from 'priceApi/price'
 import { CurrencyAmount, Currency, TradeType, Price, Percent, Fraction } from '@uniswap/sdk-core'
 import { Trade } from '@uniswap/v2-sdk'
